chore: improve module naming
