+++ src/clustree/io.py
@@ -1,29 +1,7 @@
 from pathlib import Path
 from typing import Union
 
-from PIL import Image, ImageDraw, ImageFont
-
-
-def _get_img_name_pattern(kk: int) -> list[str]:
-    """
-    Given kk, produce list of form 'K_k'. \
-    For example, kk = 2 produces ['1_1', '2_1', '2_2'].
-
-    :param kk: int, highest cluster resolution, 1 or greater.
-    :return: list, containing K_k combinations
-    """
-    return [f"{K}_{k}" for K in range(1, kk + 1) for k in range(1, K + 1)]
-
-
-def _get_fake_img(k_upper: str, k_lower: str) -> Image:
-    W, H = 40, 40
-    img = Image.new("RGB", (W, H), color=(255, 255, 255))
-    draw = ImageDraw.Draw(img)
-    font = ImageFont.load_default()
-    out = f"{k_upper}_{k_lower}"
-    _, _, w, h = draw.textbbox((0, 0), out, font=font)
-    draw.text(((W - w) / 2, (H - h) / 2), out, font=font, fill="black")
-    return img
+from PIL import Image
 
 
 def read_images(
